implementing the initial setup for the framework
